refactor: handle navigation and refactor screen contant parameter and main screen
