+++ app/src/main/java/com/baghdad/novix/di/ViewModelModule.kt
@@ -1,7 +1,9 @@
 package com.baghdad.novix.di
 
 import com.baghdad.viewmodel.actorDetails.ActorDetailsViewModel
+import com.baghdad.viewmodel.actorGallery.ActorGalleryViewModel
 import com.baghdad.viewmodel.search.SearchViewModel
+import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.androidx.viewmodel.dsl.viewModelOf
 import org.koin.dsl.module
 
@@ -9,4 +11,7 @@ val viewModelModule = module {
     viewModelOf(::SearchViewModel)
     viewModelOf(::ActorDetailsViewModel)
     viewModelOf(::ActorDetailsViewModel)
+    viewModel {
+        (actorId: Long) -> ActorGalleryViewModel(get(), actorId)
+    }
 }
