+++ ui/src/main/java/com/baghdad/ui/navigation/graph/actorDetails/ActorDetailsNavGraph.kt
@@ -4,6 +4,7 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.composable
 import androidx.navigation.navigation
+import com.baghdad.ui.feature.actorGallery.GalleryScreen
 import com.baghdad.ui.navigation.graph.DummyScreen
 import com.baghdad.ui.navigation.graph.util.toGraph
 import com.baghdad.ui.navigation.route.ActorDetailsRoute
@@ -19,7 +20,9 @@ fun NavGraphBuilder.actorDetailsNavGraph(navController: NavHostController) {
         }
         composable<ActorDetailsRoute.ActorGalleryScreen> { backStackEntry ->
             val actorId = backStackEntry.toGraph<Graph.ActorDetailsGraph>(navController).actorId
-            DummyScreen("Actor Gallery Screen $actorId")
+            GalleryScreen(actorId) { event ->
+                handleActorDetailsNavigation(event, navController)
+            }
         }
         composable<ActorDetailsRoute.ActorTopMoviePicksScreen> { backStackEntry ->
             val actorId = backStackEntry.toGraph<Graph.ActorDetailsGraph>(navController).actorId
