+++ viewmodel/src/main/java/com/baghdad/viewmodel/actorGallery/ActorGalleryInteractionListener.kt
@@ -1,0 +1,5 @@
+package com.baghdad.viewmodel.actorGallery
+
+interface ActorGalleryInteractionListener {
+    fun onBackClick()
+}
