+++ viewmodel/src/main/java/com/baghdad/viewmodel/actorGallery/ActorGalleryScreenEffect.kt
@@ -1,0 +1,8 @@
+package com.baghdad.viewmodel.actorGallery
+
+import com.baghdad.viewmodel.base.BaseUiEffect
+
+sealed class ActorGalleryScreenEffect: BaseUiEffect {
+   data object OnBackClick: ActorGalleryScreenEffect()
+
+}
