--- viewmodel/src/main/java/com/baghdad/viewmodel/gallery/GalleryInteractionListener.kt
@@ -1,5 +1,0 @@
-package com.baghdad.viewmodel.gallery
-
-interface GalleryInteractionListener {
-    fun onBackClick()
-}
