--- viewmodel/src/main/java/com/baghdad/viewmodel/gallery/GalleryScreenEffect.kt
@@ -1,8 +1,0 @@
-package com.baghdad.viewmodel.gallery
-
-import com.baghdad.viewmodel.base.BaseUiEffect
-
-sealed class GalleryScreenEffect: BaseUiEffect {
-   object OnBackClick: GalleryScreenEffect()
-
-}
